refactor auth0 lock into shared component
